+++ app/containers/App/index.js
@@ -1,7 +1,4 @@
-/**
- *
- * App
- *
+/*
  * This component is the skeleton around the actual pages, and should only
  * contain code that should be seen on all pages. (e.g. navigation bar)
  */
@@ -11,10 +8,10 @@ import { Helmet } from 'react-helmet';
 import { Switch, Route, withRouter } from 'react-router-dom';
 import Footer from 'components/Footer';
 import LoginPage from '../LoginPage';
-import FeaturePage from '../FeaturePage';
 import NotFoundPage from '../NotFoundPage/Loadable';
 import Customer from '../CustomerManagement/Customer';
 import CustomerStaff from '../CustomerManagement/CustomerStaff';
+import CustomerManagement from '../CustomerManagement';
 import Header from '../Layouts/Header';
 import Sidebar from '../Layouts/Sidebar';
 
@@ -26,15 +23,18 @@ function App() {
         titleTemplate="%s - React.js Boilerplate"
         defaultTitle="React.js Boilerplate"
       >
-        <meta name="description" content="A React.js Boilerplate application" />
+        <meta
+          name="description"
+          content="A React.js Boilerplate application"
+        />
       </Helmet>
       {/* <Header /> */}
       {/* <Sidebar /> */}
       <Switch>
         <Route path="/auth" component={LoginPage} />
-        <Route path="/features" component={FeaturePage} />
         <Route path="/customer" component={Customer} />
         <Route path="/customer-staff" component={CustomerStaff} />
+        <Route path="/customer-management" component={CustomerManagement} />
         <Route path="" component={NotFoundPage} />
       </Switch>
       {/* <Footer /> */}
