+++ package.json
@@ -81,6 +81,7 @@
     "invariant": "2.2.4",
     "ip": "1.1.5",
     "lodash": "^4.17.21",
+    "material-ui-dropzone": "^3.5.0",
     "minimist": "^1.2.5",
     "prop-types": "15.7.2",
     "react": "16.8.6",
